Move side effect code
Auth action is moved to react saga.
